requirement.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
langchain>=0.0.200
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.3.0
python-dotenv>=0.19.0